+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Adds CircleCI install generator. ([@abhaynikam][])
 * Adds GitHub Actions install generator. ([@abhaynikam][])
 * Adds Travis CI install generator. ([@abhaynikam][])
+* Adds Rubocop install generator. ([@abhaynikam][])
 
 ## 0.3.0 (September 19th, 2020)
 * Adds Bullet install generator. ([@abhaynikam][])